refactor: examples for theme 'XML'
