+++ src/main/java/com/rakovets/course/javabasics/example/xml/model/Item.java
@@ -14,27 +14,35 @@ public class Item {
     public void setDate(String date) {
         this.date = date;
     }
+
     public String getMode() {
         return mode;
     }
+
     public void setMode(String mode) {
         this.mode = mode;
     }
+
     public String getUnit() {
         return unit;
     }
+
     public void setUnit(String unit) {
         this.unit = unit;
     }
+
     public String getCurrent() {
         return current;
     }
+
     public void setCurrent(String current) {
         this.current = current;
     }
+
     public String getInteractive() {
         return interactive;
     }
+
     public void setInteractive(String interactive) {
         this.interactive = interactive;
     }
