+++ src/main/resources/example/xml/stax-reading.xml
