+++ src/main/resources/example/xml/stax-writing.xml
