+++ src/main/resources/practice/xml/hibernate.cfg.xml
